chest_xray_api/requirements.txt
fastapi
uvicorn
joblib
opencv-python
numpy
scikit-image
pillow
scikit-learn
python-multipart